requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
httplib2==0.14.0
whitenoise==5.0.1
sqlparse==0.3.0
asgiref==3.2.3